code organization updates
