+++ src/settings.js
@@ -1,4 +1,6 @@
 import { PluginSettingTab, Setting, FuzzySuggestModal, MarkdownView } from 'obsidian';
+import { createExampleSettings } from './settingsTabExample';
+import { createAPISettings } from './settingsTabAPISettings';
 
 const DEFAULT_SETTINGS = {
     apiProvider: 'all',
@@ -782,7 +784,7 @@ class PixelBannerSettingTab extends PluginSettingTab {
 
         // API Settings tab content
         const apiTab = tabContentContainer.createEl('div', { cls: 'tab-content', attr: { 'data-tab': 'API Settings' } });
-        this.createAPISettings(apiTab);
+        createAPISettings(apiTab, this.plugin);
 
         // Folder Images tab content
         const foldersTab = tabContentContainer.createEl('div', { cls: 'tab-content', attr: { 'data-tab': 'Folder Images' } });
@@ -790,7 +792,7 @@ class PixelBannerSettingTab extends PluginSettingTab {
 
         // Examples tab content
         const examplesTab = tabContentContainer.createEl('div', { cls: 'tab-content', attr: { 'data-tab': 'Examples' } });
-        this.createExampleSettings(examplesTab);
+        createExampleSettings(examplesTab, this.plugin);
 
         // Activate the General tab by default
         tabsEl.firstChild.click();
@@ -814,357 +816,6 @@ class PixelBannerSettingTab extends PluginSettingTab {
         });
 
         return { tabsEl, tabContentContainer };
-    }
-
-    createAPISettings(containerEl) {
-        // section callout
-        const calloutEl = containerEl.createEl('div', { cls: 'tab-callout' });
-        calloutEl.createEl('div', { text: 'Optionally select which API provider to use for fetching images. See the Examples tab for more information on referencing images by URL or local image. You can use any combination of API keyword, URL, or local image between notes.' });
-
-        // Add API provider radio buttons
-        new Setting(containerEl)
-            .setName('API Provider')
-            .setDesc('Select the API provider for fetching images')
-            .addDropdown(dropdown => dropdown
-                .addOption('all', 'All (Random)')
-                .addOption('pexels', 'Pexels')
-                .addOption('pixabay', 'Pixabay')
-                .addOption('flickr', 'Flickr')
-                .addOption('unsplash', 'Unsplash')
-                .setValue(this.plugin.settings.apiProvider)
-                .onChange(async (value) => {
-                    this.plugin.settings.apiProvider = value;
-                    await this.plugin.saveSettings();
-                    this.display(); // Refresh the settings tab to update API key fields
-                }));
-
-        // Pexels API key
-        new Setting(containerEl)
-            .setName('Pexels API Key');
-        containerEl.createEl('span', { text: 'Enter your Pexels API key. Get your API key from ', cls: 'setting-item-description' })
-            .createEl('a', { href: 'https://www.pexels.com/api/', text: 'Pexels API' });
-        
-        // Pexels API key setting
-        const pexelsApiKeySetting = new Setting(containerEl)
-            .setClass('full-width-control')
-            .addText(text => {
-                text
-                    .setPlaceholder('Pexels API key')
-                    .setValue(this.plugin.settings.pexelsApiKey)
-                    .onChange(async (value) => {
-                        this.plugin.settings.pexelsApiKey = value;
-                        await this.plugin.saveSettings();
-                    });
-                text.inputEl.style.width = 'calc(100% - 100px)';  // Make room for the Test button
-            })
-            .addButton(button => button
-                .setButtonText('Test API')
-                .onClick(async () => {
-                    const apiKey = this.plugin.settings.pexelsApiKey;
-                    if (!apiKey) {
-                        new Notice('Please enter an API key first');
-                        return;
-                    }
-                    
-                    button.setButtonText('Testing...');
-                    button.setDisabled(true);
-                    
-                    const isValid = await testPexelsApi(apiKey);
-                    
-                    button.setButtonText('Test API');
-                    button.setDisabled(false);
-                    
-                    new Notice(isValid ? '✅ Pexels API key is valid!' : '❌ Invalid Pexels API key');
-                }));
-        
-        pexelsApiKeySetting.settingEl.style.width = '100%';
-
-        // Pixabay API key
-        new Setting(containerEl)
-            .setName('Pixabay API Key');
-        
-        // Pixabay API key description
-        containerEl.createEl('span', { text: 'Enter your Pixabay API key. Get your API key from ', cls: 'setting-item-description' })
-            .createEl('a', { href: 'https://pixabay.com/api/docs/', text: 'Pixabay API' });
-        
-        // Pixabay API key setting
-        const pixabayApiKeySetting = new Setting(containerEl)
-            .setClass('full-width-control')
-            .addText(text => {
-                text
-                    .setPlaceholder('Pixabay API key')
-                    .setValue(this.plugin.settings.pixabayApiKey)
-                    .onChange(async (value) => {
-                        this.plugin.settings.pixabayApiKey = value;
-                        await this.plugin.saveSettings();
-                    });
-                text.inputEl.style.width = 'calc(100% - 100px)';  // Make room for the Test button
-            })
-            .addButton(button => button
-                .setButtonText('Test API')
-                .onClick(async () => {
-                    const apiKey = this.plugin.settings.pixabayApiKey;
-                    if (!apiKey) {
-                        new Notice('Please enter an API key first');
-                        return;
-                    }
-                    
-                    button.setButtonText('Testing...');
-                    button.setDisabled(true);
-                    
-                    const isValid = await testPixabayApi(apiKey);
-                    
-                    button.setButtonText('Test API');
-                    button.setDisabled(false);
-                    
-                    new Notice(isValid ? '✅ Pixabay API key is valid!' : '❌ Invalid Pixabay API key');
-                }));
-        
-        pixabayApiKeySetting.settingEl.style.width = '100%';
-
-        // Flickr API key
-        new Setting(containerEl)
-            .setName('Flickr API Key');
-        
-        // Flickr API key description
-        containerEl.createEl('span', { text: 'Enter your Flickr API key. Get your API key from ', cls: 'setting-item-description' })
-            .createEl('a', { href: 'https://www.flickr.com/services/api/', text: 'Flickr API' });
-        
-        // Flickr API key setting
-        const flickrApiKeySetting = new Setting(containerEl)
-            .setClass('full-width-control')
-            .addText(text => {
-                text
-                    .setPlaceholder('Flickr API key')
-                    .setValue(this.plugin.settings.flickrApiKey)
-                    .onChange(async (value) => {
-                        this.plugin.settings.flickrApiKey = value;
-                        await this.plugin.saveSettings();
-                    });
-                text.inputEl.style.width = 'calc(100% - 100px)';
-            })
-            .addButton(button => button
-                .setButtonText('Test API')
-                .onClick(async () => {
-                    const apiKey = this.plugin.settings.flickrApiKey;
-                    if (!apiKey) {
-                        new Notice('Please enter an API key first');
-                        return;
-                    }
-                    
-                    button.setButtonText('Testing...');
-                    button.setDisabled(true);
-                    
-                    const isValid = await testFlickrApi(apiKey);
-                    
-                    button.setButtonText('Test API');
-                    button.setDisabled(false);
-                    
-                    new Notice(isValid ? '✅ Flickr API key is valid!' : '❌ Invalid Flickr API key');
-                }));
-
-        // Unsplash API key
-        new Setting(containerEl)
-            .setName('Unsplash API Key');
-        
-        // Unsplash API key description
-        containerEl.createEl('span', { text: 'Enter your Unsplash API key (Access Key). Get your API key from ', cls: 'setting-item-description' })
-            .createEl('a', { href: 'https://unsplash.com/oauth/applications', text: 'Unsplash API' });
-        
-        // Unsplash API key setting
-        const unsplashApiKeySetting = new Setting(containerEl)
-            .setClass('full-width-control')
-            .addText(text => {
-                text
-                    .setPlaceholder('Unsplash API key')
-                    .setValue(this.plugin.settings.unsplashApiKey)
-                    .onChange(async (value) => {
-                        this.plugin.settings.unsplashApiKey = value;
-                        await this.plugin.saveSettings();
-                    });
-                text.inputEl.style.width = 'calc(100% - 100px)';
-            })
-            .addButton(button => button
-                .setButtonText('Test API')
-                .onClick(async () => {
-                    const apiKey = this.plugin.settings.unsplashApiKey;
-                    if (!apiKey) {
-                        new Notice('Please enter an API key first');
-                        return;
-                    }
-                    
-                    button.setButtonText('Testing...');
-                    button.setDisabled(true);
-                    
-                    const isValid = await testUnsplashApi(apiKey);
-                    
-                    button.setButtonText('Test API');
-                    button.setDisabled(false);
-                    
-                    new Notice(isValid ? '✅ Unsplash API key is valid!' : '❌ Invalid Unsplash API key');
-                }));
-
-        // Images section
-        new Setting(containerEl)
-            .setName('Images')
-            .setDesc('Configure settings for images fetched from API. These settings apply when using keywords to fetch random images.')
-            .setHeading();
-
-        // Show Pin Icon setting
-        new Setting(containerEl)
-            .setName('Show Pin Icon')
-            .setDesc('Show a pin icon on random banner images that allows saving them to your vault. Once pinned, your frontmatter will be updated to use the local image instead of the API image.')
-            .addToggle(toggle => toggle
-                .setValue(this.plugin.settings.showPinIcon)
-                .onChange(async (value) => {
-                    this.plugin.settings.showPinIcon = value;
-                    // Show/hide dependent settings based on the toggle
-                    folderInputSetting.settingEl.style.display = value ? 'flex' : 'none';
-                    refreshIconSetting.settingEl.style.display = value ? 'flex' : 'none';
-                    await this.plugin.saveSettings();
-                }));
-
-        // Folder input setting
-        const folderInputSetting = new Setting(containerEl)
-            .setName('Pinned Images Folder')
-            .setDesc('Default folder where pinned banner images will be saved')
-            .addText(text => {
-                text.setPlaceholder('pixel-banner-images')
-                    .setValue(this.plugin.settings.pinnedImageFolder)
-                    .onChange(async (value) => {
-                        this.plugin.settings.pinnedImageFolder = value;
-                        await this.plugin.saveSettings();
-                    });
-
-                // Add blur handler for validation
-                text.inputEl.addEventListener('blur', async (event) => {
-                    let value = text.inputEl.value.trim();
-                    
-                    if (!value) {
-                        value = 'pixel-banner-images';
-                    }
-
-                    text.setValue(value);
-                    this.plugin.settings.pinnedImageFolder = value;
-                    await this.plugin.saveSettings();
-                });
-
-                return text;
-            })
-            .addButton(button => button
-                .setButtonText('Clean Orphaned Pins')
-                .setTooltip('Remove pinned images from the default folder that are no longer referenced in Notes')
-                .onClick(async () => {
-                    button.setButtonText('🫧 Cleaning...');
-                    button.setDisabled(true);
-                    
-                    try {
-                        const result = await this.plugin.cleanOrphanedPins();
-                        new Notice(`🧼 Cleaned ${result.cleaned} orphaned pinned images`);
-                    } catch (error) {
-                        console.error('Error cleaning orphaned pins:', error);
-                        new Notice('Failed to clean orphaned pins');
-                    } finally {
-                        button.setButtonText('Clean Orphaned Pins');
-                        button.setDisabled(false);
-                    }
-                }));
-
-        // Refresh icon setting
-        const refreshIconSetting = new Setting(containerEl)
-            .setName('Show Refresh Icon')
-            .setDesc('Show a refresh icon next to the pin icon to get a new random image')
-            .addToggle(toggle => toggle
-                .setValue(this.plugin.settings.showRefreshIcon)
-                .onChange(async (value) => {
-                    this.plugin.settings.showRefreshIcon = value;
-                    await this.plugin.saveSettings();
-                }));
-
-        // Set initial visibility of dependent settings
-        folderInputSetting.settingEl.style.display = this.plugin.settings.showPinIcon ? 'flex' : 'none';
-        refreshIconSetting.settingEl.style.display = this.plugin.settings.showPinIcon ? 'flex' : 'none';
-
-        // Size setting
-        new Setting(containerEl)
-            .setName('Size')
-            .setDesc('Select the size of the image - (API only)')
-            .addDropdown(dropdown => dropdown
-                .addOption('small', 'Small')
-                .addOption('medium', 'Medium')
-                .addOption('large', 'Large')
-                .setValue(this.plugin.settings.imageSize)
-                .onChange(async (value) => {
-                    this.plugin.settings.imageSize = value;
-                    await this.plugin.saveSettings();
-                }));
-
-        // Orientation setting
-        new Setting(containerEl)
-            .setName('Orientation')
-            .setDesc('Select the orientation of the image - (API only)')
-            .addDropdown(dropdown => dropdown
-                .addOption('landscape', 'Landscape')
-                .addOption('portrait', 'Portrait')
-                .addOption('square', 'Square')
-                .setValue(this.plugin.settings.imageOrientation)
-                .onChange(async (value) => {
-                    this.plugin.settings.imageOrientation = value;
-                    await this.plugin.saveSettings();
-                }));
-
-        // Number of images setting
-        new Setting(containerEl)
-            .setName('Number of images')
-            .setDesc('Enter the number of random images to fetch (3-50) - (API only)')
-            .addText(text => text
-                .setPlaceholder('10')
-                .setValue(String(this.plugin.settings.numberOfImages || 10))
-                .onChange(async (value) => {
-                    let numValue = Number(value);
-                    if (!isNaN(numValue)) {
-                        numValue = Math.max(3, Math.min(numValue, 50)); // Ensure value is between 3 and 50
-                        this.plugin.settings.numberOfImages = numValue;
-                        await this.plugin.saveSettings();
-                    }
-                }))
-            .then(setting => {
-                const inputEl = setting.controlEl.querySelector('input');
-                inputEl.type = 'number';
-                inputEl.min = '3'; // Set minimum to 3
-                inputEl.max = '50';
-                inputEl.style.width = '50px';
-            });
-
-        // Default keywords setting
-        const defaultKeywordsSetting = new Setting(containerEl)
-            .setName('Default keywords')
-            .setDesc('Enter a comma-separated list of default keywords to be used when no keyword is provided in the frontmatter, or when the provided keyword does not return any results. - (API only)')
-            .addTextArea(text => {
-                text
-                    .setPlaceholder('Enter keywords, separated by commas')
-                    .setValue(this.plugin.settings.defaultKeywords)
-                    .onChange(async (value) => {
-                        this.plugin.settings.defaultKeywords = value;
-                        await this.plugin.saveSettings();
-                    });
-                text.inputEl.style.width = '100%';
-                text.inputEl.style.marginTop = '15px';
-                text.inputEl.style.height = '90px';
-            })
-            .addExtraButton(button => button
-                .setIcon('reset')
-                .setTooltip('Reset to default')
-                .onClick(async () => {
-                    this.plugin.settings.defaultKeywords = DEFAULT_SETTINGS.defaultKeywords;
-                    await this.plugin.saveSettings();
-                    this.display();
-                })
-            );
-
-        defaultKeywordsSetting.settingEl.dataset.id = 'defaultKeywords';
-        defaultKeywordsSetting.settingEl.style.display = 'flex';
-        defaultKeywordsSetting.settingEl.style.flexDirection = 'column';
     }
 
     createGeneralSettings(containerEl) {
@@ -2232,88 +1883,6 @@ class PixelBannerSettingTab extends PluginSettingTab {
                 }));
     }
 
-    createExampleSettings(containerEl) {
-        new Setting(containerEl)
-            .setName('How to use')
-            .setHeading()
-            .settingEl.querySelector('.setting-item-name').style.cssText = 'color: var(--text-accent-hover); font-size: var(--font-ui-large);';
-
-        // Helper function to get a random item from an array
-        const getRandomFieldName = (fieldNames) => {
-            const names = Array.isArray(fieldNames) ? fieldNames : [fieldNames];
-            return names[Math.floor(Math.random() * names.length)];
-        };
-
-        const instructionsEl = containerEl.createEl('div', { cls: 'pixel-banner-section' });
-        instructionsEl.createEl('p', { text: 'Add the following fields to your note\'s frontmatter to customize the banner:' });
-        const codeEl = instructionsEl.createEl('pre');
-        codeEl.createEl('code', { text: 
-`---
-${getRandomFieldName(this.plugin.settings.customBannerField)}: blue turtle
-${getRandomFieldName(this.plugin.settings.customYPositionField)}: 30
-${getRandomFieldName(this.plugin.settings.customXPositionField)}: 30
-${getRandomFieldName(this.plugin.settings.customContentStartField)}: 200
-${getRandomFieldName(this.plugin.settings.customImageDisplayField)}: contain
-${getRandomFieldName(this.plugin.settings.customImageRepeatField)}: true
-${getRandomFieldName(this.plugin.settings.customBannerHeightField)}: 400
-${getRandomFieldName(this.plugin.settings.customFadeField)}: -75
-${getRandomFieldName(this.plugin.settings.customBorderRadiusField)}: 25
-${getRandomFieldName(this.plugin.settings.customTitleColorField)}: #ff0000
----
-
-# Or use a direct URL:
----
-${getRandomFieldName(this.plugin.settings.customBannerField)}: https://example.com/image.jpg
-${getRandomFieldName(this.plugin.settings.customYPositionField)}: 70
-${getRandomFieldName(this.plugin.settings.customXPositionField)}: 70
-${getRandomFieldName(this.plugin.settings.customContentStartField)}: 180
-${getRandomFieldName(this.plugin.settings.customImageDisplayField)}: 200%
-${getRandomFieldName(this.plugin.settings.customBannerHeightField)}: 300
-${getRandomFieldName(this.plugin.settings.customFadeField)}: -75
-${getRandomFieldName(this.plugin.settings.customBorderRadiusField)}: 0
-${getRandomFieldName(this.plugin.settings.customTitleColorField)}: #00ff00
----
-
-# Or use a path to an image in the vault:
----
-${getRandomFieldName(this.plugin.settings.customBannerField)}: Assets/my-image.png
-${getRandomFieldName(this.plugin.settings.customYPositionField)}: 0
-${getRandomFieldName(this.plugin.settings.customXPositionField)}: 0
-${getRandomFieldName(this.plugin.settings.customContentStartField)}: 100
-${getRandomFieldName(this.plugin.settings.customImageDisplayField)}: auto
-${getRandomFieldName(this.plugin.settings.customBannerHeightField)}: 250
-${getRandomFieldName(this.plugin.settings.customFadeField)}: -75
-${getRandomFieldName(this.plugin.settings.customBorderRadiusField)}: 50
-${getRandomFieldName(this.plugin.settings.customTitleColorField)}: #0000ff
----
-
-# Or use an Obsidian internal link:
----
-${getRandomFieldName(this.plugin.settings.customBannerField)}: [[example-image.png]]
-${getRandomFieldName(this.plugin.settings.customYPositionField)}: 100
-${getRandomFieldName(this.plugin.settings.customXPositionField)}: 100
-${getRandomFieldName(this.plugin.settings.customContentStartField)}: 50
-${getRandomFieldName(this.plugin.settings.customImageDisplayField)}: contain
-${getRandomFieldName(this.plugin.settings.customImageRepeatField)}: false
-${getRandomFieldName(this.plugin.settings.customBannerHeightField)}: 500
-${getRandomFieldName(this.plugin.settings.customFadeField)}: -75
-${getRandomFieldName(this.plugin.settings.customBorderRadiusField)}: 17
-${getRandomFieldName(this.plugin.settings.customTitleColorField)}: #ff00ff
----`
-        });
-
-        instructionsEl.createEl('p', { text: 'Note: The image display options are "auto", "cover", or "contain". The image repeat option is only applicable when the display is set to "contain".' });
-
-        // Add example image
-        containerEl.createEl('img', {
-            attr: {
-                src: 'https://raw.githubusercontent.com/jparkerweb/pixel-banner/main/example.jpg',
-                alt: 'Example of a Pixel banner',
-                style: 'max-width: 100%; height: auto; margin-top: 10px; border-radius: 5px;'
-            }
-        });
-    }
-
     validateFieldName(value, otherFieldName) {
         if (value === otherFieldName) {
             new Notice("Field names must be unique!");
@@ -2368,65 +1937,6 @@ function random20characters() {
         result += characters.charAt(Math.floor(Math.random() * characters.length));
     }
     return result;
-}
-
-// API test functions
-async function testPexelsApi(apiKey) {
-    try {
-        const response = await fetch(`https://api.pexels.com/v1/search?query=${random20characters()}&per_page=3`, {
-            headers: {
-                'Authorization': apiKey
-            }
-        });
-        
-        if (!response.ok) {
-            throw new Error('❌ Invalid Pexels API key');
-        }
-        
-        const data = await response.json();
-        // console.log(`Pexels API response:`, data);
-        return data.photos; // Ensure photos are present
-    } catch (error) {
-        return false;
-    }
-}
-
-async function testPixabayApi(apiKey) {
-    try {
-        const response = await fetch(`https://pixabay.com/api/?key=${apiKey}&q=test&per_page=3`);
-        const data = await response.json();
-        
-        if (data.error) {
-            throw new Error(data.error);
-        }
-        
-        return true;
-    } catch (error) {
-        return false;
-    }
-}
-
-async function testFlickrApi(apiKey) {
-    try {
-        const response = await fetch(`https://www.flickr.com/services/rest/?method=flickr.test.echo&api_key=${apiKey}&format=json&nojsoncallback=1`);
-        const data = await response.json();
-        return data.stat === 'ok';
-    } catch (error) {
-        return false;
-    }
-}
-
-async function testUnsplashApi(apiKey) {
-    try {
-        const response = await fetch('https://api.unsplash.com/photos/random', {
-            headers: {
-                'Authorization': `Client-ID ${apiKey}`
-            }
-        });
-        return response.ok;
-    } catch (error) {
-        return false;
-    }
 }
 
 export { DEFAULT_SETTINGS, FolderSuggestModal, FolderImageSetting, PixelBannerSettingTab, debounce };
